refactor: share non-negative assertion
