+++ .changeset/shared-non-negative-assertion.md
@@ -1,0 +1,6 @@
+---
+"@morpho-org/morpho-ts": minor
+"@morpho-org/midnight-sdk": patch
+---
+
+Add shared `assertNonNegative` and `NegativeValueError` utilities to `@morpho-org/morpho-ts`, and consume them directly from `@morpho-org/midnight-sdk`.
