--- packages/midnight-sdk/src/internal.ts
@@ -1,8 +1,0 @@
-import { NegativeValueError } from "./errors.js";
-
-/**
- * @internal
- */
-export const assertNonNegative = (field: string, value: bigint) => {
-  if (value < 0n) throw new NegativeValueError(field, value);
-};
