+++ packages/morpho-ts/src/utils.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, test, vi } from "vitest";
 import {
+  assertNonNegative,
   bigIntComparator,
   createGetValue,
   createHasValue,
@@ -16,6 +17,7 @@ import {
   isNotUndefined,
   keys,
   mergeEntries,
+  NegativeValueError,
   retryPromiseLinearBackoff,
   transformValue,
   values,
@@ -476,6 +478,17 @@ describe("getLastDefined", () => {
 
   test("returns the only element when array has one defined value", () => {
     expect(getLastDefined([42])).toBe(42);
+  });
+});
+
+describe("assertNonNegative", () => {
+  test("default", () => {
+    expect(assertNonNegative("assets", 0n)).toBe(undefined);
+    expect(assertNonNegative("assets", 1n)).toBe(undefined);
+  });
+
+  test("error: NegativeValueError", () => {
+    expect(() => assertNonNegative("assets", -1n)).toThrow(NegativeValueError);
   });
 });
