+++ packages/morpho-ts/src/utils.ts
@@ -155,6 +155,40 @@ export function getLastDefined<T>(array: T[]) {
   return getLast(filterDefined(array));
 }
 
+/**
+ * Thrown when a uint-like value is negative.
+ *
+ * @example
+ * ```ts
+ * import { NegativeValueError } from "@morpho-org/morpho-ts";
+ *
+ * throw new NegativeValueError("assets", -1n);
+ * ```
+ */
+export class NegativeValueError extends Error {
+  public constructor(field: string, value: bigint) {
+    super(`${field} "${value}" must be non-negative.`);
+    this.name = "NegativeValueError";
+  }
+}
+
+/**
+ * Asserts that a bigint value is non-negative.
+ *
+ * @param field - Field name used in the thrown error message.
+ * @param value - Bigint value to validate.
+ * @throws NegativeValueError when `value` is negative.
+ * @example
+ * ```ts
+ * import { assertNonNegative } from "@morpho-org/morpho-ts";
+ *
+ * assertNonNegative("assets", 0n);
+ * ```
+ */
+export function assertNonNegative(field: string, value: bigint) {
+  if (value < 0n) throw new NegativeValueError(field, value);
+}
+
 export function deepFreeze<T>(obj: T): T {
   if (obj === null || obj === undefined) {
     // null or undefined are already immutable
